+++ tests/plot.R
@@ -109,3 +109,7 @@ plot(nc, pal = rainbow, breaks = (0:10)/3)
 # logz:
 nc$e = 10^(nc$SID74)
 plot(nc["e"], logz = TRUE)
+
+# shared key:
+plot(nc[c("SID74", "SID79")], key.pos = -1)
+plot(nc[c("BIR74", "BIR79")], key.pos = 1, logz=TRUE)